# be/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Clinic Management API",
        "title": "Clinic Management",
        "contact": {},
        "version": "1.0"
    },
    "host": "clinicly.fly.dev",
    "basePath": "/api/v1",
    "paths": {
...
        "/medical-report": {
            "get": {
                "description": "Get medical report",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medical report"
                ],
                "summary": "Get medical report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient",
                        "name": "patient",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Order by",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Order descending",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medical report response",
                        "schema": {
                            "$ref": "#/definitions/controllers.MedicalReportListResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create medical report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medical report"
                ],
                "summary": "Create medical report",
                "parameters": [
                    {
                        "description": "Medical report data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.MedicalReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medical report response",
                        "schema": {
                            "$ref": "#/definitions/controllers.MedicalReportResponse"
                        }
                    }
                }
            }
        },
        "/medical-report/{id}": {
            "get": {
                "description": "Get medical report by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medical report"
                ],
                "summary": "Get medical report by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medical report response",
                        "schema": {
                            "$ref": "#/definitions/controllers.MedicalReportResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update medical report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medical report"
                ],
                "summary": "Update medical report",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Medical report id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Medical report data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateMedicalReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medical report response",
                        "schema": {
                            "$ref": "#/definitions/controllers.MedicalReportResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete medical report",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medical report"
                ],
                "summary": "Delete medical report",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Medical report id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medical report response",
                        "schema": {
                            "$ref": "#/definitions/controllers.MedicalReportResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.MedicalReportListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MedicalReport"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "controllers.MedicalReportRequest": {
            "type": "object",
            "required": [
                "doctor",
                "patient"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "diagnose": {
                    "type": "string"
                },
                "doctor": {
                    "type": "integer"
                },
                "patient": {
                    "type": "integer"
                },
                "prescription": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.PrescriptionRequest"
                    }
                }
            }
        },
        "controllers.MedicalReportResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.MedicalReport"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "controllers.PrescriptionRequest": {
            "type": "object",
            "required": [
                "instruction",
                "medicine",
                "quantity"
            ],
            "properties": {
                "instruction": {
                    "type": "string"
                },
                "medicine": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "controllers.UpdateMedicalReportRequest": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "diagnose": {
                    "type": "string"
                },
                "doctor": {
                    "type": "integer"
                },
                "patient": {
                    "type": "integer"
                },
                "prescription": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.PrescriptionRequest"
                    }
                }
            }
        },
...
        "models.MedicalReport": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "diagnose": {
                    "type": "string"
                },
                "doctor": {
                    "$ref": "#/definitions/models.Staff"
                },
                "doctor_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "patient_id": {
                    "type": "integer"
                },
                "prescription": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Prescription"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
...
        "models.Prescription": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "instruction": {
                    "type": "string"
                },
                "medical_report_id": {
                    "type": "integer"
                },
                "medicine_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
...
        "models.Staff": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "identity_card": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "salary": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
...
    }
}